requirements.txt
fastmcp
pydantic
supabase
google-analytics-data
google-auth
requests
python-dotenv
